introduced EthereumAdapter
