+++ subscription/test/subscription.spec.js
@@ -1,18 +1,6 @@
 const {expect} = require("chai");
 const Orbs = require("orbs-client-sdk");
-const { deploySubscriptionManager } = require("./truffle-scripts/deploySubscriptionManager");
-const { getCurrentBlock } = require("./truffle-scripts/getCurrentBlock");
-const { mine } = require("./truffle-scripts/mine");
-
-
-//TODO this currently only handles Ganache - for ropsten or mainnet we need to add busywaits with sleeps
-async function waitForBlock(blockNumber) {
-    const currentBlock = await getCurrentBlock();
-    const blocksToMine = blockNumber - currentBlock.number;
-    if (blocksToMine) {
-        await mine(blocksToMine);
-    }
-}
+const {EthereumAdapter} = require("./ethereumAdapter");
 
 function sleep(ms) {
     return new Promise(resolve => {
@@ -21,6 +9,14 @@ function sleep(ms) {
 }
 
 describe("orbs network", async () => {
+    let ethereum;
+    before(async ()=>{
+        ethereum = await EthereumAdapter.build();
+    });
+
+    after(()=>{
+        ethereum.stop();
+    });
 
     it("rejects transactions after refreshing when subscription in not valid", async () => {
 
@@ -34,12 +30,12 @@ describe("orbs network", async () => {
 
         expect(result1.executionResult).to.equal("SUCCESS");
 
-        const subscriptionManager = await deploySubscriptionManager();
+        const subscriptionManager = await ethereum.deploySubscriptionManager();
         expect(subscriptionManager.address).to.be.ok;
 
         // finality
-        const currentBlock = await getCurrentBlock();
-        await waitForBlock(currentBlock.number + 10);
+        const currentBlock = await ethereum.getLatestBlock();
+        await ethereum.waitForBlock(currentBlock.number + 10);
         await sleep(2000);
 
         const [refreshSubTx] = client.createTransaction(signer.publicKey, signer.privateKey, "_GlobalPreOrder", "refreshSubscription", [Orbs.argString(subscriptionManager.address)]);
@@ -51,7 +47,6 @@ describe("orbs network", async () => {
         const result2 = await client.sendTransaction(t2);
 
         expect(result2.executionResult).to.equal("NOT_EXECUTED");
-
     })
 });
